move impl to convert.rs
